Report_Agent/requirements.txt
pydantic>=2.0.0
requests
beautifulsoup4
sentence-transformers
faiss-cpu
openai>=0.28.0
nltk
rich
python-dotenv
streamlit
pytest
flake8
